int input

+++ src/RoboticArm6DOF/Kinematics.h
@@ -7,8 +7,8 @@
 namespace lra {
 	class Kinematics {
 	public:
-		static const glm::vec3& GetForwardKinematics(const JointAngles& joint_angles, const JointLength& lra_dimensions);
-		static lra::JointAngles GetInverseKinematics(bool& is_valid, const glm::vec3& target, const JointLength& lra_dimensions);
-		static lra::JointAngles GetInverseKinematics(bool &is_valid, const glm::vec3& target, const JointLength& lra_dimensions, float phi);
+		static const glm::ivec3& GetForwardKinematics(const JointAngles& joint_angles, const JointLength& lra_dimensions);
+		static lra::JointAngles GetInverseKinematics(bool& is_valid, const glm::ivec3& target, const JointLength& lra_dimensions);
+		static lra::JointAngles GetInverseKinematics(bool &is_valid, const glm::ivec3& target, const JointLength& lra_dimensions, float phi);
 	};
 }
